Revert "modify function addpicture"
This reverts commit 452ae940a5e4af83381f574f865e2ef62a2bdd95.

+++ templates/functions.php
@@ -206,7 +206,7 @@ try {
 }
 }
 
-function addPicture($picture,$file_name){
+function addPicture($picture){
 	// $file = array();
 	// foreach ($picture as $key1 => $value1) {
 	// 	foreach ($value1 as $key2 => $value2) {
@@ -224,6 +224,7 @@ $error="";
 	 $allowedExts = array("jpg", "jpeg", "gif", "png", "bmp");
 				$tmp_extension = explode(".", $file["name"]);
 				$extension = end($tmp_extension);
+
 				if (
 						(
 							 ($file["type"] == "image/gif")
@@ -234,22 +235,39 @@ $error="";
 						&& ($file["size"] < 2000000)
 						&& in_array($extension, $allowedExts))
 					{
+
+
 				 if ($file["error"] > 0)
 								{
+
 										$error.= "Return Code: " . $file["error"] . "<br />";
+
 								} else {
+
 										$error.= "Upload: " . $file["name"] . "<br />";
 										$error.= "Type: " . $file["type"] . "<br />";
 										$error.=  "Size: " . ($file["size"] / 1024) . " Kb<br />";
 										$error.= "Temp file: " . $file["tmp_name"] . "<br />";
-										move_uploaded_file($file["tmp_name"],
-										"upload/" . $_SESSION["username"] . "." . $extension);
-										$error.= "Stored in: " . "upload/" . $file_name . "." . $extension;
+
+										if (file_exists("upload/" . $file["name"])) {
+											$error.= $file["name"] . " already exists. ";
+
+										} else {
+											move_uploaded_file($file["tmp_name"],
+											"upload/" . $_SESSION["username"] . "." . $extension);
+											$error.= "Stored in: " . "upload/" . $_SESSION["username"] . "." . $extension;
+										}
 								}
+
 					}    else {
 
 						$error.= $file["type"]."<br />";
 							$error.= "Invalid file try another Image";
+
 					}
 }
+
+
+
+
 ?>
